[add] Create docs about "mod_inv()" and "mod_pow()"

+++ src/Math/ModularArithmetic/mod_pow.hpp
@@ -1,17 +1,18 @@
-/**
- * @brief 繰り返し二乗法（mod付き）
- */
-
 #ifndef ALGORITHM_MOD_POW_HPP
 #define ALGORITHM_MOD_POW_HPP 1
+
+/**
+ * @brief 繰り返し二乗法（mod付き）
+ * @docs docs/Math/ModularArithmetic/mod_pow.md
+ */
 
 #include <cassert>
 
 namespace algorithm {
 
 // 繰り返し二乗法（mod付き）．
-// a^k (mod m) を求める．O(logK).
-long long mod_pow(long long a, long long k, int m) {
+// a^k mod m を求める．O(logK).
+int mod_pow(long long a, long long k, int m) {
     assert(k >= 0);
     assert(m >= 1);
     long long res = 1;
